requirements.txt
# dependecies required for python versions 3.8 - 3.12
build==1.2.1; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.8'
charset-normalizer==3.3.2; python_version >= '3.8'
idna==3.7; python_version >= '3.8'
packaging==24.0; python_version >= '3.8'
pyproject_hooks==1.1.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.8'
# dependecies required for testing (due to vcrpy)
multidict==6.0.5; python_version >= '3.8'
PyYAML==6.0.1; python_version >= '3.8'
vcrpy==6.0.1; python_version >= '3.8'
wrapt==1.16.0; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.8'

# dependecies required for python versions 3.8 - 3.10
importlib_metadata==7.1.0; python_version == '3.8' or python_version == '3.9'
tomli==2.0.1; python_version == '3.8' or python_version == '3.9' or python_version == '3.10'
zipp==3.18.1; python_version == '3.8' or python_version == '3.9'

# other dependecies
urllib3==2.2.1; python_version >= '3.10'
urllib3==1.26.18; python_version == '3.8' or python_version == '3.9' # conflicting with vcrpy on these versions
